+++ CHANGELOG.md
@@ -9,13 +9,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- [tut pkg] Added some more test util functions
+- [tut pkg] `CookieCheckFunc` type for cookie checks
+
 ### Changed
 
 - Upgraded github.com/onsi/gomega to v1.19.0 to get rid of go mod tidy issues
+- [tut pkg] `RespHasSetCookie` now takes `CookieCheckFunc` as checks
 
 ### Deprecated
 
 ### Removed
+
+- [tut pkg] Removed `RespBodyEqualsText` and `RespBodyTextMatchesRegex`. `RespHasTextBody` should be used instead
 
 ### Fixed